Support removing the package dependency list from an existing package
This change allows callers to change the dependencies of a package
to a blank list. On doing so, any pre-existing dependencies list in
the yml file will be deleted - rather than preserved.

Detail:

When writing a new yml file, the #write_package_yml! method tries to
keep the existing yml config intact so far as possible, so as to
preserve custom fields/data in the yml file.

The `data` attribute on ParsePackwerk::Package contains the original
config, and the attributes supplied as parameters to ParsePackwerk::Package.new
contain the new values that the config _should_ be.

The existing implementation worked fine for other attributes, but the
`dependencies` check used `package.dependencies.any?`

Since the `any?` method returns false if the array is blank, this meant that
any attempt to overwrite the existing package list with a blank list kept the
original items intact... rather than deleting the clause from the yml file

This change allows the caller to supply either `dependencies: nil`
OR `dependencies: []` and the output file that `ParsePackwerk.write_package_yml!`
creates will then no it's `dependencies` clause removed entirely.